Complete refactor with tests and validators (#5)
* Removed Data Annotations from DTOs

* Added tests for FileQuery and HolidaysQueries

* Refactors and GetAllRequestsForUserByDateQueryHandler Tests

* Expanded CircleCI config

* fix: circleci config fixed

* Updated config.yml

* fix: modified circleci config to use docker

* fix: typo fixed

* Updated config.yml

* Added .editorconfig file

* Added GetAllRequestsQueryHandler tests

* fix: refactors and added GetAllRequestsForUserQueryHandler tests

* Added more tests for GetAllRequestsForUserQueryHandler

* Added validator for AddNewHolidayCommand

* Added Holiday and RequestValidators and renewed exception handling

* updated circleci config

* Added validation for every command

* Replaced ApplicationException with BusinessException

* Change IDs to Guid

* Fixed GetFileByFileNameQueryHandlerTests

+++ WorkSchedule.Application/CommandHandlers/Holidays/DeleteHolidayCommandHandler.cs
@@ -1,11 +1,7 @@
 ﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using WorkSchedule.Api.Commands.Holidays;
 using WorkSchedule.Application.Data;
+using WorkSchedule.Application.Exceptions;
 
 namespace WorkSchedule.Application.CommandHandlers.Holidays
 {
@@ -20,7 +16,13 @@ namespace WorkSchedule.Application.CommandHandlers.Holidays
 
         public async Task<Unit> Handle(DeleteHolidayCommand request, CancellationToken cancellationToken)
         {
-            var holiday = _uow.HolidayRepository.GetByID(request.Id);
+            var validator = new DeleteHolidayCommanddValidator();
+            var validatorResult = await validator.ValidateAsync(request, cancellationToken);
+            if (!validatorResult.IsValid)
+            {
+                throw new BusinessException { ErrorCode = 599, ErrorMessages = validatorResult.Errors.Select(e => e.ErrorMessage).ToList() };
+            }
+            var holiday = await _uow.HolidayRepository.GetByID(request.Id);
             _uow.HolidayRepository.Delete(holiday);
             _uow.Save();
             return Unit.Value;
